requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
paddlepaddle==2.6.2
paddleocr==2.9.1
pdf2image>=1.16.0
deep-translator>=1.11.0
indic-transliteration>=2.3.0
requests>=2.31.0
python-dotenv>=1.0.0
